requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
airsim==1.8.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==24.3.0
backports.ssl_match_hostname==3.7.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
Django==4.2.8
django-extensions==3.2.3
django-restframework==0.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.6
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langchain-upstage==0.4.0
langsmith==0.2.7
matplotlib==3.9.1
matplotlib-inline==0.1.7
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
oauthlib==3.2.2
openai==1.58.1
opencv-contrib-python==4.10.0.84
orjson==3.10.11
packaging==24.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==11.0.0
pinecone==5.4.2
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.10.0
pyparsing==3.1.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
tornado==4.5.3
tqdm==4.67.0
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.18.3
